Introduce `JUICEFS_IMMUTABLE`, mount `/etc/updatedb.conf` only in mutable environments
In immutable environments, such as Talos Linux, most host directories are read-only.
As a result, for example, `FileOrCreate` for the `/etc/updatedb.conf` will fail to
create an empty file. This patch introduces `JUICEFS_IMMUTABLE` as a new environment
variable for both controller and node instances to handle these kinds of cases that
need special attention in immutable environments.